+++ client/src/containers/user-form/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { UserFormProps } from './types';