+++ client/src/containers/user-form/types.ts
@@ -1,0 +1,6 @@
+import type { ProfilePayload } from 'types';
+
+export type UserFormProps = {
+  user?: ProfilePayload;
+  closeUserForm?: () => void;
+};