+++ client/src/layouts/profile/constants.ts
@@ -1,3 +1,5 @@
+import { RoleName } from 'hooks/permissions/enums';
+
 import type { TabsType } from 'components/tabs';
 
 export const USER_TABS: TabsType = {
@@ -8,5 +10,6 @@ export const USER_TABS: TabsType = {
   USERS: {
     name: 'Users',
     href: '/profile/users',
+    restrictedToRoles: [RoleName.ADMIN],
   },
 };