+++ client/src/pages/profile/users.tsx
@@ -2,6 +2,8 @@ import { useMemo, useState } from 'react';
 import Head from 'next/head';
 import { useDebounceCallback } from '@react-hook/debounce';
 import { PlusIcon } from '@heroicons/react/solid';
+import { capitalize } from 'lodash-es';
+import Image from 'next/image';
 
 import { useUsers } from 'hooks/users';
 import ProfileLayout from 'layouts/profile';
@@ -9,10 +11,15 @@ import Button from 'components/button';
 import Search from 'components/search';
 import Table from 'components/table';
 import { DEFAULT_PAGE_SIZES } from 'components/table/pagination/constants';
+import { usePermissions } from 'hooks/permissions';
+import { RoleName } from 'hooks/permissions/enums';
+import Modal from 'components/modal/component';
+import UserForm from 'containers/user-form';
+import { NEW_USER } from 'containers/user-form/helpers';
 
 import type { PaginationState, SortingState } from '@tanstack/react-table';
 import type { TableProps } from 'components/table/component';
-import type { User } from 'types';
+import type { ProfilePayload } from 'types';
 
 const AdminUsersPage: React.FC = () => {
   const [search, setSearch] = useState<string>('');
@@ -21,31 +28,81 @@ const AdminUsersPage: React.FC = () => {
     pageIndex: 1,
     pageSize: DEFAULT_PAGE_SIZES[0],
   });
+  const [selectedUser, setOpenUserFormModal] = useState<ProfilePayload>();
+
+  const { hasRole } = usePermissions();
+  const isAdmin = hasRole(RoleName.ADMIN);
 
   const sortStr = useMemo(() => {
     return sorting.map((sort) => (sort.desc ? `-${sort.id}` : sort.id)).join(',') || null;
   }, [sorting]);
 
-  const { data, isFetching } = useUsers({
-    'page[size]': pagination.pageSize,
-    'page[number]': pagination.pageIndex,
-    sort: sortStr,
-  });
+  const { data, isFetching } = useUsers(
+    {
+      'page[size]': pagination.pageSize,
+      'page[number]': pagination.pageIndex,
+      sort: sortStr,
+    },
+    {
+      enabled: isAdmin,
+    },
+  );
 
   const handleOnSearch = useDebounceCallback((searchTerm: string) => setSearch(searchTerm), 250);
 
-  const tableProps = useMemo<TableProps<User>>(
+  const tableProps = useMemo<TableProps<ProfilePayload>>(
     () => ({
       state: { pagination, sorting },
       columns: [
-        { id: 'displayName', header: 'Name', size: 110, align: 'left', enableSorting: true },
-        { id: 'email', header: 'Email', align: 'left', enableSorting: true },
         {
-          id: 'isActive',
-          header: 'Active',
-          cell: ({ row }) => (row.original.isActive ? 'Yes' : 'No'),
+          id: 'fname',
+          header: 'Name',
           align: 'left',
           enableSorting: true,
+          cell: ({ row }) => (
+            <div className="flex items-center gap-4" data-testname="user-table-row">
+              <div className="flex-shrink-0 w-10 h-10 rounded-full">
+                <Image
+                  width={40}
+                  height={40}
+                  src={row.original.avatarDataUrl || '/images/avatar.png'}
+                  alt="User image"
+                />
+              </div>
+              <span>{`${row.original.fname || ''} ${row.original.lname || ''}`}</span>
+            </div>
+          ),
+        },
+        { id: 'email', header: 'Email', align: 'left', enableSorting: true },
+        {
+          id: 'displayName',
+          header: 'Title',
+          align: 'left',
+          enableSorting: true,
+        },
+        {
+          id: 'roles',
+          header: 'Role',
+          cell: ({ row }) => row.original.roles.map((role) => capitalize(role.name)),
+          align: 'center',
+          size: 150,
+          enableSorting: false,
+        },
+        {
+          id: 'edit',
+          header: '',
+          size: 75,
+          cell: ({ row }) => (
+            <Button
+              variant="white"
+              size="xs"
+              onClick={() => setOpenUserFormModal(row.original)}
+              disabled={!isAdmin}
+              className="h-8"
+            >
+              Edit
+            </Button>
+          ),
         },
       ],
       data: data?.data ?? [],
@@ -60,8 +117,20 @@ const AdminUsersPage: React.FC = () => {
         totalItems: data?.meta?.totalItems,
       },
     }),
-    [data, pagination, sorting],
+    [data, pagination, sorting, isAdmin],
   );
+
+  const modalTitle = useMemo(() => {
+    if (selectedUser?.id) {
+      if (selectedUser.fname || selectedUser.lname) {
+        return `Edit user "${selectedUser.fname || ''}${
+          selectedUser.lname ? ` ${selectedUser.lname}` : ''
+        }"`;
+      }
+      return 'Edit user';
+    }
+    return 'Add user';
+  }, [selectedUser]);
 
   return (
     <ProfileLayout title="Users">
@@ -84,12 +153,21 @@ const AdminUsersPage: React.FC = () => {
               </div>
             }
             disabled
+            // disabled={!isAdmin} uncomment when endpoint ready
+            onClick={() => setOpenUserFormModal(NEW_USER)}
           >
             Add user
           </Button>
         </div>
       </div>
       <Table {...tableProps} isLoading={isFetching} />
+      <Modal
+        title={modalTitle}
+        open={!!selectedUser}
+        onDismiss={() => setOpenUserFormModal(undefined)}
+      >
+        <UserForm user={selectedUser} closeUserForm={() => setOpenUserFormModal(undefined)} />
+      </Modal>
     </ProfileLayout>
   );
 };